Cache the collection ID lookup
Formerly, with the `id_lists` variable being a module-level variable,
this leads to the unexpected behavior of making a bunch of requests even
if all you do is import this module.

By slapping a cache onto the `get_ids()` function, it will only issue
the requests the first time it's called and then reuse the results for
all the subsequent calls.