Use NDArray[Any] instead of ArrayLike
ArrayLike makes sense for an input type but np.asarray() or similar
should probably be called on it to ensure an ndarray is present. But
ArrayLike probably isn't suitable as an output type. I'm not sure what
the dtype for the arrays all are, so the weakest one is Any.